refactor: 채팅 카드 위젯 컴포넌트화
