+++ analysis_options.yaml
@@ -10,8 +10,10 @@
 analyzer:
   errors:
     prefer_const_constructors: ignore
+    prefer_const_constructors_in_immutables: ignore
     prefer_const_literals_to_create_immutables: ignore
     sized_box_for_whitespace: ignore
+    use_key_in_widget_constructors: ignore
 include: package:flutter_lints/flutter.yaml
 
 linter:
