+++ lib/main.dart
@@ -2,7 +2,7 @@ import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:moreyourbong/firebase_options.dart';
-import 'package:moreyourbong/views/pages/chat_page.dart';
+import 'package:moreyourbong/views/pages/chat/chat_page.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
