+++ lib/views/pages/chat/chat_page.dart
@@ -1,0 +1,93 @@
+import 'dart:math';
+
+import 'package:flutter/material.dart';
+import 'package:moreyourbong/views/pages/chat/widgets/chatting_card.dart';
+
+class ChatPage extends StatelessWidget {
+  const ChatPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    double bottomSheetHeight = 100;
+    return GestureDetector(
+      onTap: () {
+        FocusScope.of(context).unfocus();
+      },
+      child: Scaffold(
+        backgroundColor: Colors.grey[100],
+        appBar: AppBar(
+          backgroundColor: Color(0xFFF8F4E8),
+          centerTitle: true,
+          title: Text(
+            "모임 이름",
+            style: TextStyle(fontWeight: FontWeight.bold),
+          ),
+        ),
+        body: ListView.separated(
+          padding: EdgeInsets.fromLTRB(12, 20, 12, bottomSheetHeight + 10),
+          itemCount: 7,
+          separatorBuilder: (context, index) {
+            return SizedBox(height: 10);
+          },
+          itemBuilder: (context, index) {
+            return tempDataList[index];
+          },
+        ),
+        bottomSheet: SafeArea(
+          child: Container(
+            padding: EdgeInsets.symmetric(horizontal: 16, vertical: 12),
+            width: double.infinity,
+            decoration: BoxDecoration(
+              color: Colors.white,
+              border: Border(
+                top: BorderSide(
+                  width: 1,
+                  color: Colors.grey[300]!,
+                ),
+              ),
+            ),
+            child: Row(
+              children: [
+                Expanded(
+                  child: TextField(
+                    decoration: InputDecoration(
+                      enabledBorder: OutlineInputBorder(
+                        borderRadius: BorderRadius.circular(24),
+                        borderSide: BorderSide(color: Colors.grey[500]!),
+                      ),
+                      focusedBorder: OutlineInputBorder(
+                        borderRadius: BorderRadius.circular(24),
+                        borderSide: BorderSide(
+                          width: 2,
+                          color: Color(0xff4CAF50),
+                        ),
+                      ),
+                    ),
+                  ),
+                ),
+                SizedBox(width: 8),
+                GestureDetector(
+                  onTap: () {
+                    print("send tap");
+                  },
+                  child: Container(
+                    width: 50,
+                    height: 50,
+                    decoration: BoxDecoration(
+                      color: Color(0xFF4CAF50),
+                      borderRadius: BorderRadius.circular(50),
+                    ),
+                    child: Icon(
+                      Icons.send,
+                      color: Colors.white,
+                    ),
+                  ),
+                ),
+              ],
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+}
